extracted to a utils fn
